Showing the Current Conditions in the tab
