+++ src/app/current-conditions/current-conditions.component.html
@@ -1,7 +1,6 @@
 <div>
   <div class="well flex" (click)="showForecast(location.zip)">
     <div>
-      <h3>{{location.data.name}} ({{location.zip}})</h3>
       <h4>Current conditions: {{location.data.weather[0].main}}</h4>
       <h4>Temperatures today:</h4>
       <p>
@@ -14,21 +13,7 @@
       </p>
     </div>
     <div>
-      <span class="close" (click)="removeLocation(location.zip)">&times;</span>
       <img [src]="location.data.weatherIcon">
     </div>
  </div>
-</div>
-
-<div>
-    <app-tabs-container>
-        <app-tabs>
-            <app-tab [tabId]="'1'" [title]="'Tab 1'">
-                The tab 1 content goes here ...
-            </app-tab>
-            <app-tab [tabId]="'2'" [title]="'Tab 2'">
-                The tab 2 content goes here ...
-            </app-tab>
-        </app-tabs>
-    </app-tabs-container>
 </div>
