--- src/app/tabs-engine/tabs/tabs.component.css
@@ -1,36 +1,0 @@
-.tab-header {
-    display: inline-flex;
-    align-items: center;
-    gap: 15px;
-    cursor: pointer;
-    height: 100%;
-    border: 1px solid black;
-    padding: 4px 10px;
-    background-color: #627b9a;
-    color: white;
-    margin: 1px 1px;
-}
-
-.tabs-content {
-    padding: 20px 15px;
-    height: 100%;
-    border: 1px solid black;
-}
-
-.close-btn {
-    width: 20px;
-    height: 20px;
-    padding: 3px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &:hover {
-        background-color: rgba(0, 0, 0, 0.1);
-        cursor: pointer;
-    }
-
-    img {
-        width: 100%;
-    }
-}
